Revert "Update windows build instructions" - this breaks packaging on Linux.
This reverts commit c6d9d22ebb797b6da853c996d3ddd20cb1f772a6.

+++ modules/camera/src/libifm3d_camera/camera_impl.hpp
@@ -883,8 +883,7 @@ ifm3d::Camera::Impl::SetCurrentTime(int epoch_seconds)
 {
   if (epoch_seconds < 0)
     {
-      epoch_seconds = static_cast<int>(
-                std::chrono::seconds(std::time(nullptr)).count());
+      epoch_seconds = std::chrono::seconds(std::time(nullptr)).count();
     }
 
   this->_XCallTime("setCurrentTime", epoch_seconds);
